+++ changelog.md
@@ -2,6 +2,7 @@
 
 ### Latest
 
+- Adding duplicated jets plot to after resampling plots [!714](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/714)
 - Improve resampled file flavour label handling [!713](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/713)
 - Make track weight variable customizable [!712](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/712)
 - Add support for different labels for each track-like group [!688](https://gitlab.cern.ch/atlas-flavor-tagging-tools/algorithms/umami/-/merge_requests/688)